requirements.txt
gradio
requests
pillow
Flask
